Moved interface ExportsWebComponent to internal package, removed getPropertyConfigurationMap from interface
